requirements.txt
openai
google-generativeai
python-dateutil
rich
pydantic
requests
